+++ src/Language/Qux/Annotated/Parser.hs
@@ -70,15 +70,14 @@ decl = choice [functionDecl, importDecl] <?> "declaration"
         functionDecl    = do
             pos <- getPosition
 
+            attrs <- many functionAttribute
             name <- id_
             symbol_ "::"
             parameterTypes <- (try $ (,) <$> type_ <*> id_) `endBy` rightArrow
             returnType <- type_
-            colon
-            indented
-            stmts <- block stmt
+            stmts <- option [] (colon >> indented >> block stmt)
 
-            return $ FunctionDecl pos name (parameterTypes ++ [(returnType, Id pos "@")]) stmts
+            return $ FunctionDecl pos attrs name (parameterTypes ++ [(returnType, Id pos "@")]) stmts
         importDecl      = do
             pos <- getPosition
 
@@ -86,6 +85,10 @@ decl = choice [functionDecl, importDecl] <?> "declaration"
             id <- sepBy1 id_ dot
 
             return $ ImportDecl pos id
+
+-- | 'FunctionAttribute' parser.
+functionAttribute :: Parser (FunctionAttribute SourcePos)
+functionAttribute = getPosition >>= \pos -> External pos <$ reserved "external"
 
 -- | 'Stmt' parser.
 stmt :: Parser (Stmt SourcePos)